+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Unreleased
+
+- Use bytes.IndexByte instead of Contains
+
+
 ## v0.0.5
 
 ### Improvements